Move diffable
